+++ src/test/java/edu/illinois/library/cantaloupe/ApplicationTest.java
@@ -31,7 +31,7 @@ public class ApplicationTest extends TestCase {
             Path fixturePath = Paths.get(cwd, "src", "test", "resources");
             config.setProperty("print_stack_trace_on_error_pages", false);
             config.setProperty("http.port", PORT);
-            config.setProperty("processor.fallback", "ImageIoProcessor");
+            config.setProperty("processor.fallback", "Java2dProcessor");
             config.setProperty("resolver", "FilesystemResolver");
             config.setProperty("FilesystemResolver.path_prefix", fixturePath +
                     File.separator);